requirements.txt
typer
docker
jinja2
pyyaml 